requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
python-dotenv
slowapi